+++ exceptions/customExceptions.ts
@@ -1,7 +1,3 @@
-class PasswordsNotMatched extends Error {
-	constructor(code: number, message: string) {
-		super(message);
-	}
-}
-
-export default PasswordsNotMatched;
+/**
+ * Loading...
+ */