+++ routes/vote.route.ts
@@ -5,6 +5,6 @@ import vote from '../controllers/vote.controller';
 
 const router = Router();
 
-router.post('/', authorize, vote.addVote);
+router.post('/', authorize, vote.updateVote);
 
 export default router;